requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-rest-knox==4.2.0
django-cors-headers==4.4.0
requests==2.32.3
Pillow==10.4.0
python-decouple==3.8